test: More test for DB. Empty DB before and after test. (#18)

+++ team2_project/src/test/java/org/csed332/project/team2/DBTest.java
@@ -4,26 +4,61 @@ import org.csed332.project.team2.db.model.MetricModel;
 import org.csed332.project.team2.db.service.MetricModelService;
 import org.csed332.project.team2.db.util.HibernateUtil;
 import org.hibernate.Session;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.*;
+
+import java.util.List;
 
 public class DBTest {
-    @BeforeEach
-    void emptyDB() {
+    private MetricModel getExampleMetricModel() {
+        MetricModel m = new MetricModel();
+        m.setClassName("ClassA");
+        m.setMetric("MetricA");
+        m.setFigure(1.0);
+        return m;
+    }
+
+    @BeforeAll
+    public static void emptyDBBefore() {
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             session.beginTransaction();
-            session.createQuery("delete from MetricModel ").executeUpdate();
+            List l = session.createQuery("from MetricModel ").list();
+            if (!l.isEmpty()) {
+                session.createQuery("delete from MetricModel ").executeUpdate();
+            }
             session.getTransaction().commit();
         }
     }
 
     @Test
-    void testDBSaveAndGet() {
-        MetricModel m = new MetricModel();
-        m.setClassName("ClassA");
-        m.setMetric("MetricA");
-        m.setFigure(1.0);
+    public void testDBGetNull() {
+        MetricModel m = MetricModelService.getMetricById(1234L);
+        Assertions.assertNull(m);
+    }
+
+    @Test
+    public void testDBSave() {
+        MetricModel m = getExampleMetricModel();
+        MetricModelService.save(m);
+
+        List l;
+        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+            session.beginTransaction();
+            l = session.createQuery("from MetricModel ").list();
+            session.getTransaction().commit();
+        }
+
+        Assertions.assertEquals(1, l.size());
+
+        MetricModel mNew = (MetricModel)l.get(0);
+        Assertions.assertEquals(m.getCreated().getTime(), mNew.getCreated().getTime());
+        Assertions.assertEquals("ClassA", mNew.getClassName());
+        Assertions.assertEquals("MetricA", mNew.getMetric());
+        Assertions.assertEquals(1.0, mNew.getFigure());
+    }
+
+    @Test
+    public void testDBSaveAndGet() {
+        MetricModel m = getExampleMetricModel();
 
         MetricModelService.save(m);
 
@@ -32,7 +67,31 @@ public class DBTest {
         Assertions.assertEquals("ClassA", mNew.getClassName());
         Assertions.assertEquals("MetricA", mNew.getMetric());
         Assertions.assertEquals(1.0, mNew.getFigure());
+    }
 
-        MetricModelService.remove(mNew);
+    @Test
+    public void testDBSaveAndRemove() {
+        MetricModel m = getExampleMetricModel();
+
+        MetricModelService.save(m);
+
+        MetricModelService.remove(m);
+
+        List l;
+        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+            session.beginTransaction();
+            l = session.createQuery("from MetricModel ").list();
+            session.getTransaction().commit();
+        }
+        Assertions.assertTrue(l.isEmpty());
+    }
+
+    @AfterEach
+    public void emptyDBAfter() {
+        try (Session session = HibernateUtil.getSessionFactory().openSession()) {
+            session.beginTransaction();
+            session.createQuery("delete from MetricModel ").executeUpdate();
+            session.getTransaction().commit();
+        }
     }
 }
